Fix timezone issues with shiftinfo times
Setting the time zone in the initializer didn't seem to have had any effect. Strange.
So when we set the time zone in the application.rb starting with Rails 6.1, all the times changed. Gah.
Furthermore, since the time in the database were not "naïve", daylight savings time would have an ill effect. (All the more since the one-hour shift would occur right during the planning phase for the current year.)
So back to UTC. Bern would be nicer but meh.